TP 1 Agencia de Viaje

+++ TP_1/src/Menu.h
@@ -4,9 +4,9 @@
 
 /** \brief Imprime las opciones a elegir del menú
  *
- * \param kilometros float Es la cantidad de kilómetros de Vuelo
- * \param precioAero float Es el precio de Aerolíneas
- * \param precioLatam float Es el precio de Latam
+ * \param kilometros double Es la cantidad de kilómetros de Vuelo
+ * \param precioAero double Es el precio de Aerolíneas
+ * \param precioLatam double Es el precio de Latam
  * \param flagKm int Bandera que verifica si los kilómetros están cargados
  * \param flagAero int Bandera que verifica si el precio de Aerolíneas está cargado
  * \param flagLatam int Bandera que verifica si el precio de Latam está cargado
@@ -15,81 +15,81 @@
  * \return int 0 si está todo Ok o -1 si hubo un error
  *
  */
-int mostrarMenuOpciones(float kilometros, float precioAero, float precioLatam, int flagKm, int flagAero, int flagLatam, int* opcion);
+int mostrarMenuOpciones(double kilometros, double precioAero, double precioLatam, int flagKm, int flagAero, int flagLatam, int* opcion);
 
 
 /** \brief Realiza los cálculos de los costos de Vuelo
  *
- * \param kilometros float Es la cantidad de kilómetros de Vuelo
- * \param precio float Es el precio de Vuelo
- * \param tarDeb float* Lugar donde se guardará el cálculo de tarjeta de débito de Vuelo
- * \param tarCred float* Lugar donde se guardará el cálculo de tarjeta de crédito de Vuelo
+ * \param kilometros double Es la cantidad de kilómetros de Vuelo
+ * \param precio double Es el precio de Vuelo
+ * \param tarDeb double* Lugar donde se guardará el cálculo de tarjeta de débito de Vuelo
+ * \param tarCred double* Lugar donde se guardará el cálculo de tarjeta de crédito de Vuelo
  * \param bitcoin double* Lugar donde se guardará el cálculo de bitcoin de Vuelo
- * \param unitario float* Lugar donde se guardará el cálculo de precio unitario de Vuelo
+ * \param unitario double* Lugar donde se guardará el cálculo de precio unitario de Vuelo
  *
  * \return int 0 si está todo Ok o -1 si hubo un error
  *
  */
-int calcularCostosVuelo(float kilometros, float precio, float* tarDeb, float* tarCred, double* bitcoin, float* unitario);
+int calcularCostosVuelo(double kilometros, double precio, double* tarDeb, double* tarCred, double* bitcoin, double* unitario);
 
 
 /** \brief Realiza los cálculos de todos los costos de Aerolíneas y Latam
  *
- * \param kilometros float Es la cantidad de kilómetros de Vuelo
- * \param precioAero float Es el precio de Aerolíneas
- * \param tarDebAero float* Lugar donde se guardará el cálculo de tarjeta de débito de Aerolíneas
- * \param tarCredAero float* Lugar donde se guardará el cálculo de tarjeta de crédito de Aerolíneas
+ * \param kilometros double Es la cantidad de kilómetros de Vuelo
+ * \param precioAero double Es el precio de Aerolíneas
+ * \param tarDebAero double* Lugar donde se guardará el cálculo de tarjeta de débito de Aerolíneas
+ * \param tarCredAero double* Lugar donde se guardará el cálculo de tarjeta de crédito de Aerolíneas
  * \param bitcoinAero double* Lugar donde se guardará el cálculo de bitcoin de Aerolíneas
- * \param unitarioAero float* Lugar donde se guardará el cálculo de precio unitario de Aerolíneas
- * \param precioLatam float Es el precio de Latam
- * \param tarDebLatam float* Lugar donde se guardará el cálculo de tarjeta de débito de Latam
- * \param tarCredLatam float* Lugar donde se guardará el cálculo de tarjeta de crédito de Latam
+ * \param unitarioAero double* Lugar donde se guardará el cálculo de precio unitario de Aerolíneas
+ * \param precioLatam double Es el precio de Latam
+ * \param tarDebLatam double* Lugar donde se guardará el cálculo de tarjeta de débito de Latam
+ * \param tarCredLatam double* Lugar donde se guardará el cálculo de tarjeta de crédito de Latam
  * \param bitcoinLatam double* Lugar donde se guardará el cálculo de bitcoin de Latam
- * \param unitarioLatam float* Lugar donde se guardará el cálculo de precio unitario de Latam
- * \param difPrecios float* Lugar donde se guardará el cálculo de diferencia de precios entre Latam y Aerolíneas
+ * \param unitarioLatam double* Lugar donde se guardará el cálculo de precio unitario de Latam
+ * \param difPrecios double* Lugar donde se guardará el cálculo de diferencia de precios entre Latam y Aerolíneas
  *
  * \return int 0 si está todo Ok o -1 si hubo un error
  *
  */
-int calcularCostosTodos(float kilometros, float precioAero, float* tarDebAero, float* tarCredAero, double* bitcoinAero, float* unitarioAero,
-                        float precioLatam, float* tarDebLatam, float* tarCredLatam, double* bitcoinLatam, float* unitarioLatam, float* difPrecios);
+int calcularCostosTodos(double kilometros, double precioAero, double* tarDebAero, double* tarCredAero, double* bitcoinAero, double* unitarioAero,
+                        double precioLatam, double* tarDebLatam, double* tarCredLatam, double* bitcoinLatam, double* unitarioLatam, double* difPrecios);
 
 
 /** \brief Imprime los resultados de los cálculos de costos de Vuelo
  *
  * \param nombre char* Es el nombre de Vuelo
- * \param precio float Es el precio de Vuelo
- * \param tarDeb float Es el resultado del cálculo de tarjeta de débito de Vuelo
- * \param tarCred float Es el resultado del cálculo de tarjeta de crédito de Vuelo
+ * \param precio double Es el precio de Vuelo
+ * \param tarDeb double Es el resultado del cálculo de tarjeta de débito de Vuelo
+ * \param tarCred double Es el resultado del cálculo de tarjeta de crédito de Vuelo
  * \param bitcoin double Es el resultado del cálculo de bitcoin de Vuelo
- * \param unitario float Es el resultado del cálculo de precio unitario de Vuelo
+ * \param unitario double Es el resultado del cálculo de precio unitario de Vuelo
  *
  * \return int 0 si está todo Ok o -1 si hubo un error
  *
  */
-int mostrarCalculosVuelo(char* nombre, float precio, float tarDeb, float tarCred, double bitcoin, float unitario);
+int mostrarCalculosVuelo(char* nombre, double precio, double tarDeb, double tarCred, double bitcoin, double unitario);
 
 
 /** \brief Imprime los resultados de los cálculos de todos los costos de Aerolíneas y Latam
  *
- * \param kilometros float Es la cantidad de kilómetros de Vuelo
- * \param precioAero float Es el precio de Aerolíneas
- * \param tarDebAero float Es el resultado del cálculo de tarjeta de débito de Aerolíneas
- * \param tarCredAero float Es el resultado del cálculo de tarjeta de crédito de Aerolíneas
+ * \param kilometros double Es la cantidad de kilómetros de Vuelo
+ * \param precioAero double Es el precio de Aerolíneas
+ * \param tarDebAero double Es el resultado del cálculo de tarjeta de débito de Aerolíneas
+ * \param tarCredAero double Es el resultado del cálculo de tarjeta de crédito de Aerolíneas
  * \param bitcoinAero double Es el resultado del cálculo de bitcoin de Aerolíneas
- * \param unitarioAero float Es el resultado del cálculo de precio unitario de Aerolíneas
- * \param precioLatam float Es el precio de Latam
- * \param tarDebLatam float Es el resultado del cálculo de tarjeta de débito de Latam
- * \param tarCredLatam float Es el resultado del cálculo de tarjeta de crédito de Latam
+ * \param unitarioAero double Es el resultado del cálculo de precio unitario de Aerolíneas
+ * \param precioLatam double Es el precio de Latam
+ * \param tarDebLatam double Es el resultado del cálculo de tarjeta de débito de Latam
+ * \param tarCredLatam double Es el resultado del cálculo de tarjeta de crédito de Latam
  * \param bitcoinLatam double Es el resultado del cálculo de bitcoin de Latam
- * \param unitarioLatam float Es el resultado del cálculo de precio unitario de Latam
- * \param difPrecios float Es el resultado del cálculo de diferencia de precios entre Latam y Aerolíneas
+ * \param unitarioLatam double Es el resultado del cálculo de precio unitario de Latam
+ * \param difPrecios double Es el resultado del cálculo de diferencia de precios entre Latam y Aerolíneas
  *
  * \return int 0 si está todo Ok o -1 si hubo un error
  *
  */
-int informarResultados(float kilometros, float precioAero, float tarDebAero, float tarCredAero, double bitcoinAero, float unitarioAero,
-		               float precioLatam, float tarDebLatam, float tarCredLatam, double bitcoinLatam, float unitarioLatam, float difPrecios);
+int informarResultados(double kilometros, double precioAero, double tarDebAero, double tarCredAero, double bitcoinAero, double unitarioAero,
+		               double precioLatam, double tarDebLatam, double tarCredLatam, double bitcoinLatam, double unitarioLatam, double difPrecios);
 
 
 /** \brief Realiza una carga forzada de datos, calcula todos los costos y
